+++ lib/locationProvider.dart
@@ -1,0 +1,6 @@
+/// Location data provider identifier, either GPS based fine location data
+/// or Network based coarse location data, can be requested
+class LocationProvider {
+  static const String Network = 'network';
+  static const String GPS = 'gps';
+}
