+++ lib/locationServiceProvider.dart
@@ -1,0 +1,7 @@
+/// How to get Location data from system, 
+/// either using android location manager
+/// or using google mobile service i.e. GMS based location manager
+class LocationServiceProvider {
+  static const String GMSBasedLocation = '0';
+  static const String LocationManagerBasedLocation = '1';
+}
